Add files via upload
In this Python project, we used Pandas library to analyze the data about banking. The information was in a file called 'bank-full.csv'.

In this project we are using various functions such as we upload the data and checked the first few rows of the data to get an idea of what it looks like then we used head(), columns, info, and dtypes.

After that, we made some pictures (graphs)to help understand the data better:
One picture showed how many people subscribed the deposit.
Another picture, called a heatmap, helped us see if certain pieces of information in the data are connected.

Zip files:-
The graphs were saved as 'graph.png' and we put them into separate zip file'Bankfull.zip' and data save another zip file 'Bankdata.zip'. 

PPT FILE:-
We make the ppt also where we have mentioned all our analyzation and mentioned all the details.

Word Documents:-
we also make the word documents file, where we mentioned end to end points.

Conclusions:-
This way, anyone can easily download and check our analysis. It's like putting everything in a digital folder for others to use or learn from.